+++ R/utilities.R
@@ -1,40 +1,20 @@
-#' Convert mean and sd to log mean for a log normal distribution
+#' Numerically stable convolution function for two pmf vectors
 #'
-#' @description `r lifecycle::badge("stable")`
-#' Convert from mean and standard deviation to the log mean of the
-#' lognormal distribution. Useful for defining distributions supported by
-#' `estimate_infections()`, `epinow()`, and `regional_epinow()`.
-#' @param mean Numeric, mean of a distribution
-#' @param sd Numeric, standard deviation of a distribution
+#' Unlike [stats::convolve()], this function does not use the FFT algorithm,
+#' which can generate negative numbers when below machine precision.
 #'
-#' @return The log mean of a lognormal distribution
+#' @param a Numeric vector, the first sequence.
+#' @param b Numeric vector, the second sequence.
+#' @return A numeric vector representing the convolution of `a` and `b`.
 #' @keywords internal
-#'
-#' @examples
-#' \dontrun{
-#' convert_to_logmean(2, 1)
-#'
-convert_to_logmean <- function(mean, sd) {
-  log(mean^2 / sqrt(sd^2 + mean^2))
+stable_convolve <- function(a, b) {
+  n <- length(a)
+  m <- length(b)
+  result <- numeric(n + m - 1)
+  for (i in seq_along(a)) {
+    for (j in seq_along(b)) {
+      result[i + j - 1] <- result[i + j - 1] + a[i] * b[m - j + 1]
+    }
+  }
+  result
 }
-
-#' Convert mean and sd to log standard deviation for a log normal distribution
-#'
-#' @description `r lifecycle::badge("stable")`
-#' Convert from mean and standard deviation to the log standard deviation of the
-#' lognormal distribution. Useful for defining distributions supported by
-#' `estimate_infections()`, `epinow()`, and `regional_epinow()`.
-#' @param mean Numeric, mean of a distribution
-#' @param sd Numeric, standard deviation of a distribution
-#'
-#' @return The log standard deviation of a lognormal distribution
-#' @keywords internal
-#'
-#' @examples
-#' \dontrun{
-#' convert_to_logsd(2, 1)
-#' }
-convert_to_logsd <- function(mean, sd) {
-  sqrt(log(1 + (sd^2 / mean^2)))
-}
-